move to Dataset method
