+++ doc/source/ray-air/doc_code/checkpoint_usage.py
@@ -24,7 +24,6 @@ checkpoint = TensorflowCheckpoint.from_model(model)
 
 # __use_trainer_checkpoint_start__
 import ray
-from ray.data import train_test_split
 from ray.train.xgboost import XGBoostTrainer
 from ray.air.config import ScalingConfig
 
@@ -32,7 +31,7 @@ from ray.air.config import ScalingConfig
 dataset = ray.data.read_csv("s3://anonymous@air-example-data/breast_cancer.csv")
 
 # Split data into train and validation.
-train_dataset, valid_dataset = train_test_split(dataset, test_size=0.3)
+train_dataset, valid_dataset = dataset.train_test_split(test_size=0.3)
 
 trainer = XGBoostTrainer(
     scaling_config=ScalingConfig(num_workers=2),
