+++ doc/source/ray-air/examples/tf_tabular_starter.py
@@ -4,14 +4,13 @@
 # __air_generic_preprocess_start__
 import ray
 from ray.data.preprocessors import StandardScaler
-from ray.data import train_test_split
 from ray.air.config import ScalingConfig
 
 
 dataset = ray.data.read_csv("s3://anonymous@air-example-data/breast_cancer.csv")
 
 # Split data into train and validation.
-train_dataset, valid_dataset = train_test_split(dataset, test_size=0.3)
+train_dataset, valid_dataset = dataset.train_test_split(test_size=0.3)
 
 # Create a test dataset by dropping the target column.
 test_dataset = valid_dataset.map_batches(
