+++ doc/source/ray-air/examples/xgboost_batch_prediction.py
@@ -1,13 +1,12 @@
 import ray
 from ray.data.preprocessors import StandardScaler
-from ray.data import train_test_split
 from ray.train.batch_predictor import BatchPredictor
 from ray.train.xgboost import XGBoostTrainer, XGBoostPredictor
 from ray.air.config import ScalingConfig
 
 # Split data into train and validation.
 dataset = ray.data.read_csv("s3://anonymous@air-example-data/breast_cancer.csv")
-train_dataset, valid_dataset = train_test_split(dataset, test_size=0.3)
+train_dataset, valid_dataset = dataset.train_test_split(test_size=0.3)
 test_dataset = valid_dataset.drop_columns(["target"])
 
 columns_to_scale = ["mean radius", "mean texture"]
