+++ python/ray/air/BUILD
@@ -196,14 +196,6 @@ py_test(
 )
 
 py_test(
-    name = "test_dataset_utils",
-    size = "small",
-    srcs = ["tests/test_dataset_utils.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"]
-)
-
-py_test(
     name = "test_keras_callback",
     size = "small",
     srcs = ["tests/test_keras_callback.py"],
